requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coverage==5.5
Django==3.2.6
django-filter==21.1
djangorestframework==3.12.4
drf-extensions==0.7.1
Faker==8.14.0
iniconfig==1.1.1
kombu==5.1.0
packaging==21.0
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
transliterate==1.10.2
vine==5.0.0
wcwidth==0.2.5
